requirements.txt
fastapi
uvicorn
opencv-python
numpy
requests
ultralytics
pillow
python-multipart